+++ CHANGELOG.md
@@ -18,7 +18,9 @@ All notable changes to this project will be documented in this file, in reverse 
 
 ### Fixed
 
-  [#44](https://github.com/phly/http/pull/44) Retain case sensitivity for
+- [#41](https://github.com/phly/http/pull/41) Fixed examples in README to
+  reflect updates in API.
+- [#44](https://github.com/phly/http/pull/44) Retain case sensitivity for
   header names (when returned via `getHeaders()`.
 - [#45](https://github.com/phly/http/pull/45) Fix port numeric comparison to
   properly accept port integers specified via strings.